macOS icon and refactoring
